Re-structuring to use the shared steps and CommitDocumentObject
