+++ src/GitModel.AcceptanceTests/GitModel.AcceptanceTests.csproj
@@ -83,6 +83,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Steps\CommitFileReaderSteps.cs" />
     <Compile Include="Steps\CommitFileWriterSteps.cs" />
+    <Compile Include="Steps\CommitSharedSteps.cs" />
     <Compile Include="Steps\Keys.cs" />
     <Compile Include="Transformations\StringTransformations.cs" />
   </ItemGroup>
