+++ src/GitModel.AcceptanceTests/Steps/CommitSharedSteps.cs
@@ -1,0 +1,39 @@
+using TechTalk.SpecFlow;
+using FluentAssertions;
+using GitModel.AcceptanceTests.ModelObjects;
+
+namespace GitModel.AcceptanceTests.Steps
+{
+   [Binding]
+   public class CommitSharedSteps
+   {
+      private readonly CommitDocumentObject _commitDocumentObject;
+
+      public CommitSharedSteps( CommitDocumentObject commitDocumentObject )
+      {
+         _commitDocumentObject = commitDocumentObject;
+      }
+
+      [Given( @"the commit subject is ""(.*)""" )]
+      public void GivenTheCommitDocumentSubjectIs( string subject )
+      {
+         _commitDocumentObject.Subject = subject;
+      }
+
+      [Then( @"the commit file subject is ""(.*)""" )]
+      public void ThenTheCommitFileSubjectIs( string expectedSubject )
+      {
+         _commitDocumentObject.Load();
+
+         _commitDocumentObject.Subject.Should().Be( expectedSubject );
+      }
+
+      [Then( @"the commit file subject body has the lines ""(.*)""" )]
+      public void ThenTheCommitFileSubjectBodyHasTheLines( string[] bodyLines )
+      {
+         _commitDocumentObject.Load();
+
+         _commitDocumentObject.Body.Should().Contain( bodyLines );
+      }
+   }
+}
